remove all files not referred to java version of task, refactor the trashy code
It seems that some of the module test methods are useless(e.g. AirportTest.getExperimentalPlanesHasNoClassificationLevelUnclassified) because they are not testing direct functionality of class, but i've decided to not delete them entirely since i'm not sure are they a part of initial task or just a trash code. Same about sorting methods of Airport class (e.g. sortPlanesByMaxDistance), it's seems more convenient and flexibly for me to use Stream API instead so the methods looks a kind of trashy for me, but if that operation is intended to be used very frequently in the future and more flexible sorting won't be needed then they have a right for existence